+++ gatsby/src/styles/_settings.scss
@@ -2,5 +2,7 @@ $nds-colour-hero-background: #fff;
 $nds-colour-text-inverse: #222;
 $nds-container-max-width: 1170;
 $nds-container-gutter: core.$nds-spacing-medium;
+
+/*TODO remove the above and obtain them from the correct locations*/
 $bnf-brand-colour: #93da49;
 $bnfc-brand-colour: #ef4a81;
